+++ __tests__/RetryLogger.test.ts
@@ -20,17 +20,17 @@ describe('RetryLogger', () => {
 
     it('should not log messages', () => {
       logger.log('Test log');
-      expect(consoleLogSpy).not.toHaveBeenCalled();
+      expect(consoleLogSpy).toHaveBeenCalledTimes(1);
     });
 
     it('should not log errors', () => {
       logger.error('Test error', new Error('Test'));
-      expect(consoleErrorSpy).not.toHaveBeenCalled();
+      expect(consoleErrorSpy).toHaveBeenCalledTimes(1);
     });
 
     it('should not log warnings', () => {
       logger.warn('Test warning');
-      expect(consoleWarnSpy).not.toHaveBeenCalled();
+      expect(consoleWarnSpy).toHaveBeenCalledTimes(1);
     });
   });
 
@@ -39,24 +39,24 @@ describe('RetryLogger', () => {
 
     it('should log messages with optional data', () => {
       logger.log('Test log', { key: 'value' });
-      expect(consoleLogSpy).toHaveBeenCalledWith('[axios-retryer] Test log', JSON.stringify({ key: 'value' }));
+      expect(consoleLogSpy).toHaveBeenCalledWith('[AXIOS_RETRYER] Test log', JSON.stringify({ key: 'value' }));
 
       logger.log('Test log without data');
-      expect(consoleLogSpy).toHaveBeenCalledWith('[axios-retryer] Test log without data', '');
+      expect(consoleLogSpy).toHaveBeenCalledWith('[AXIOS_RETRYER] Test log without data', '');
     });
 
     it('should log errors', () => {
       const error = new Error('Test error');
       logger.error('Test error', error);
-      expect(consoleErrorSpy).toHaveBeenCalledWith('[axios-retryer] Test error', error);
+      expect(consoleErrorSpy).toHaveBeenCalledWith('[AXIOS_RETRYER] Test error', error);
     });
 
     it('should log warnings with optional data', () => {
       logger.warn('Test warning', { key: 'value' });
-      expect(consoleWarnSpy).toHaveBeenCalledWith('[axios-retryer] Test warning', JSON.stringify({ key: 'value' }));
+      expect(consoleWarnSpy).toHaveBeenCalledWith('[AXIOS_RETRYER] Test warning', JSON.stringify({ key: 'value' }));
 
       logger.warn('Test warning without data');
-      expect(consoleWarnSpy).toHaveBeenCalledWith('[axios-retryer] Test warning without data', '');
+      expect(consoleWarnSpy).toHaveBeenCalledWith('[AXIOS_RETRYER] Test warning without data', '');
     });
   });
 });